Move SourceLocation to type_definitions
