+++ src/eve/__init__.py
@@ -34,7 +34,6 @@ from .concepts import (
     FrozenNode,
     Model,
     Node,
-    SourceLocation,
     VType,
     field,
     in_field,
@@ -56,6 +55,7 @@ from .type_definitions import (
     NegativeInt,
     PositiveFloat,
     PositiveInt,
+    SourceLocation,
     Str,
     StrEnum,
     StrictBool,
